Давление заголовков и функций: void transposeSquareMatrix(matrix m);
position getMinValuePos(matrix m);

position getMaxValuePos(matrix m);

matrix createMatrixFromArray(const int *a, int nRows, int nCols);

matrix *createArrayOfMatrixFromArray(const int *values, size_t nMatrices, size_t nRows, size_t nCols);

+++ includes/structs/matrix.c
@@ -1,3 +1,5 @@
+#include <stdio.h>
+
 typedef struct matrix {
     int **values;  // элементы матриц
     int nRows;     // кол - во строк
@@ -5,6 +7,6 @@ typedef struct matrix {
 } matrix;
 
 typedef struct position {
-    int rowIndex;
-    int colIndex;
+    size_t rowIndex;
+    size_t colIndex;
 } position;
